Main/GuestSession: build fix (please review).
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@42702 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/include/GuestSessionImpl.h
@@ -70,7 +70,7 @@ public:
                       const Utf8Str &strSource, const Utf8Str &strDest, uint32_t uFlags);
 
     SessionTaskCopyTo(GuestSession *pSession,
-                      PRTFILE pSourceFile, size_t cbSourceOffset, size_t cbSourceSize,
+                      PRTFILE pSourceFile, size_t cbSourceOffset, uint64_t cbSourceSize,
                       const Utf8Str &strDest, uint32_t uFlags);
 
     virtual ~SessionTaskCopyTo(void);
